refactor: refactor outbox
