+++ ecommerce/src/main/kotlin/io/hhplus/ecommerce/config/event/EventRegistry.kt
@@ -1,12 +1,7 @@
-package io.hhplus.ecommerce.common.outbox
+package io.hhplus.ecommerce.config.event
 
 /**
  * 이벤트 레지스트리 - 모든 도메인 이벤트를 중앙에서 관리
- *
- * 목적:
- * - 이벤트 타입의 명시적 정의
- * - 이벤트 흐름의 가시성 확보
- * - 이벤트 핸들러 매핑의 중앙화
  *
  * Saga 패턴 흐름:
  * ┌─────────────────────────────────────────────────────────────────────────────┐
@@ -46,23 +41,9 @@ package io.hhplus.ecommerce.common.outbox
  * │  ─────────────────────────────────────────────────────────────────────────  │
  * │  [PaymentFailed] ──────► OrderEventHandler (주문 상태 → PAYMENT_FAILED)      │
  * └─────────────────────────────────────────────────────────────────────────────┘
- *
- * 사용법:
- * 1. 새로운 이벤트 추가: EventTypes에 상수 정의
- * 2. 핸들러 등록: EventHandler 구현 및 @Component 등록
- * 3. 이벤트 발행: OutboxEventService.publishEvent() 사용
  */
 object EventRegistry {
 
-    /**
-     * 이벤트 타입 정의
-     *
-     * 형식: {Domain}{Action}
-     * - Order 도메인: OrderCreated, OrderCancelled, OrderConfirmed
-     * - Payment 도메인: PaymentCompleted, PaymentFailed
-     * - Inventory 도메인: StockDeducted, StockRestored (내부 이벤트)
-     * - Coupon 도메인: CouponUsed, CouponRestored (내부 이벤트)
-     */
     object EventTypes {
         // Order 도메인 이벤트 (Saga 시작점)
         const val ORDER_CREATED = "OrderCreated"
@@ -86,9 +67,6 @@ object EventRegistry {
         const val POINT_REFUNDED = "PointRefunded"
     }
 
-    /**
-     * 애그리게이트 타입 정의
-     */
     object AggregateTypes {
         const val ORDER = "Order"
         const val PAYMENT = "Payment"
@@ -99,10 +77,6 @@ object EventRegistry {
         const val CART = "Cart"
     }
 
-    /**
-     * 이벤트 메타데이터
-     * - 이벤트 타입별 설명, 발행자, 구독자 정보
-     */
     data class EventMetadata(
         val eventType: String,
         val description: String,
@@ -110,17 +84,7 @@ object EventRegistry {
         val subscribers: List<String>
     )
 
-    /**
-     * 이벤트 카탈로그 - 모든 이벤트의 메타데이터
-     *
-     * Saga 흐름에 따른 이벤트 발행/구독 관계:
-     * - OrderCreated: 주문 생성 → 결제 처리 트리거
-     * - PaymentCompleted: 결제 완료 → 재고/포인트/쿠폰/배송/장바구니 처리
-     * - PaymentFailed: 결제 실패 → 주문 상태 변경
-     * - OrderCancelled: 주문 취소 → 재고 복구, 포인트 환불
-     */
     val catalog: Map<String, EventMetadata> = mapOf(
-        // === Saga Step 1: 주문 생성 ===
         EventTypes.ORDER_CREATED to EventMetadata(
             eventType = EventTypes.ORDER_CREATED,
             description = "주문이 생성되었을 때 발행. PaymentEventHandler가 결제를 처리함",
@@ -128,23 +92,21 @@ object EventRegistry {
             subscribers = listOf("PaymentEventHandler")
         ),
 
-        // === Saga Step 2: 결제 완료 (분기점) ===
         EventTypes.PAYMENT_COMPLETED to EventMetadata(
             eventType = EventTypes.PAYMENT_COMPLETED,
             description = "결제가 완료되었을 때 발행. 포인트 차감, 재고 차감, 쿠폰 사용, 배송 생성, 장바구니 정리, 판매 랭킹 업데이트, 데이터 플랫폼 전송을 병렬로 트리거함",
             publisher = "PaymentEventHandler",
             subscribers = listOf(
-                "PointEventHandler",          // 포인트 차감 (포인트 도메인 책임)
-                "InventoryEventHandler",      // 재고 차감
-                "CouponEventHandler",         // 쿠폰 사용 처리
-                "DeliveryEventHandler",       // 배송 생성
-                "CartEventHandler",           // 장바구니 정리
-                "ProductRankingEventHandler", // 판매 랭킹 업데이트 (Redis Sorted Set)
-                "OrderDataPlatformHandler"    // 데이터 플랫폼 전송 (Kafka)
+                "PointEventHandler",
+                "InventoryEventHandler",
+                "CouponEventHandler",
+                "DeliveryEventHandler",
+                "CartEventHandler",
+                "ProductRankingEventHandler",
+                "OrderDataPlatformHandler"
             )
         ),
 
-        // === 결제 실패 (보상 트랜잭션 트리거) ===
         EventTypes.PAYMENT_FAILED to EventMetadata(
             eventType = EventTypes.PAYMENT_FAILED,
             description = "결제가 실패했을 때 발행. 주문 상태를 PAYMENT_FAILED로 변경",
@@ -152,18 +114,16 @@ object EventRegistry {
             subscribers = listOf("OrderEventHandler")
         ),
 
-        // === 주문 취소 (보상 트랜잭션) ===
         EventTypes.ORDER_CANCELLED to EventMetadata(
             eventType = EventTypes.ORDER_CANCELLED,
             description = "주문이 취소되었을 때 발행. 재고 복구, 포인트 환불을 트리거함",
             publisher = "OrderCommandUseCase",
             subscribers = listOf(
-                "InventoryEventHandler",  // 재고 복구
-                "PointEventHandler"        // 포인트 환불
+                "InventoryEventHandler",
+                "PointEventHandler"
             )
         ),
 
-        // === 주문 확정 ===
         EventTypes.ORDER_CONFIRMED to EventMetadata(
             eventType = EventTypes.ORDER_CONFIRMED,
             description = "주문이 확정되었을 때 발행. 통계 기록 등 후처리를 트리거함",
@@ -172,16 +132,10 @@ object EventRegistry {
         )
     )
 
-    /**
-     * 이벤트 타입의 유효성 검증
-     */
     fun isValidEventType(eventType: String): Boolean {
         return catalog.containsKey(eventType)
     }
 
-    /**
-     * 이벤트의 구독자 목록 조회
-     */
     fun getSubscribers(eventType: String): List<String> {
         return catalog[eventType]?.subscribers ?: emptyList()
     }
