+++ ecommerce/src/main/kotlin/io/hhplus/ecommerce/order/application/handler/OrderDataPlatformHandler.kt
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import io.hhplus.ecommerce.order.application.mapper.toOrderInfoPayload
 import io.hhplus.ecommerce.common.messaging.MessagePublisher
 import io.hhplus.ecommerce.common.outbox.EventHandler
-import io.hhplus.ecommerce.common.outbox.EventRegistry
+import io.hhplus.ecommerce.config.event.EventRegistry
 import io.hhplus.ecommerce.common.outbox.OutboxEvent
 import io.hhplus.ecommerce.order.application.usecase.GetOrderQueryUseCase
 import mu.KotlinLogging
